# Conflicts:
#	src/main/java/com/snafu/todss/sig/sessies/application/SpecialInterestGroupService.java
#	src/main/java/com/snafu/todss/sig/sessies/presentation/controller/SpecialInterestGroupController.java

+++ src/main/java/com/snafu/todss/sig/sessies/application/SpecialInterestGroupService.java
@@ -62,5 +62,9 @@ public class SpecialInterestGroupService {
     }
 
     public void deleteSpecialInterestGroup(UUID id) {
+    public void deleteSpecialInterestGroup(UUID id) throws NotFoundException {
+        if (!this.SIG_REPOSITORY.existsById(id)){
+            throw new NotFoundException("No special interest group found with given id");
+        }
         this.SIG_REPOSITORY.deleteById(id);
     }}
